fix(input): fix missing keyboard events
On the title screen specifically, there was an annoying situation where keyboard events would get swallowed and not do anything. This was due to the `keyQueueReadPos` being incremented not only in `pollKey`, but also in `GameShell.run()`.

I'm not sure why the key queue read index gets increased in the main 'run' loop as well as in pollKey, but by removing it from the main loop, keyboard inputs are no longer disappearing when OnDemand is churning away on the title login screen. It could be due to browser lag messing with the async timings.